refactor: rename runner retry rate limit tests
